requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
python-dotenv
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
psycopg[binary]
python-multipart
email-validator
openpyxl
aiosmtplib
apscheduler
cryptography

